# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API para sistema de gestão de serviços
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Materials
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order with its items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/service-items:
    post:
      operationId: addOrderServiceItem
      tags: [orders]
      summary: Add a service item to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderServiceItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderServiceItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/service-items/{itemId}:
    delete:
      operationId: deleteOrderServiceItem
      tags: [orders]
      summary: Remove a service item from an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/material-items:
    post:
      operationId: addOrderMaterialItem
      tags: [orders]
      summary: Add a material item to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderMaterialItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderMaterialItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/material-items/{itemId}:
    delete:
      operationId: deleteOrderMaterialItem
      tags: [orders]
      summary: Remove a material item from an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Summary
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Client
...
    Service:
      type: object
      required: [id, name, unitPrice, unit, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        unit:
          type: string
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required: [name, unitPrice, unit]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        unitPrice:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        unitPrice:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1

    # Material
...
    Order:
      type: object
      required: [id, clientId, clientName, title, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        totalPrice:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderDetail:
      type: object
      required: [id, clientId, clientName, title, status, totalPrice, createdAt, serviceItems, materialItems]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        totalPrice:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        serviceItems:
          type: array
          items:
            $ref: "#/components/schemas/OrderServiceItem"
        materialItems:
          type: array
          items:
            $ref: "#/components/schemas/OrderMaterialItem"
    OrderInput:
      type: object
      required: [clientId, title]
      properties:
        clientId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
    OrderUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
    OrderServiceItem:
      type: object
      required: [id, orderId, serviceId, serviceName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
    OrderServiceItemInput:
      type: object
      required: [serviceId, quantity]
      properties:
        serviceId:
          type: integer
        quantity:
          type: number
          minimum: 0.01
        unitPrice:
          type: number
          minimum: 0
    OrderMaterialItem:
      type: object
      required: [id, orderId, materialId, materialName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        materialId:
          type: integer
        materialName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        totalPrice:
          type: number
    OrderMaterialItemInput:
      type: object
      required: [materialId, quantity]
      properties:
        materialId:
          type: integer
        quantity:
          type: number
          minimum: 0.01
        unitPrice:
          type: number
          minimum: 0

    # Summary
...