# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  clientId: number;
  clientName: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: OrderStatus;
  totalPrice: number;
  createdAt: Date;
  updatedAt?: Date;
}