# lib/api-zod/src/generated/types/orderDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDetailStatus } from './orderDetailStatus';
import type { OrderMaterialItem } from './orderMaterialItem';
import type { OrderServiceItem } from './orderServiceItem';

export interface OrderDetail {
  id: number;
  clientId: number;
  clientName: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: OrderDetailStatus;
  totalPrice: number;
  createdAt: Date;
  updatedAt?: Date;
  serviceItems: OrderServiceItem[];
  materialItems: OrderMaterialItem[];
}