# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputStatus } from './orderInputStatus';

export interface OrderInput {
  clientId: number;
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: OrderInputStatus;
}