# lib/api-zod/src/generated/types/orderMaterialItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */

export interface OrderMaterialItem {
  id: number;
  orderId: number;
  materialId: number;
  materialName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}