# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API para sistema de gestão de serviços
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  unitPrice: number;
  unit: string;
  createdAt: Date;
}